fix: derive team level from XP instead of stale DB value
After the XP decimal precision migration, teams.level was not
recalculated, causing wrong level display (Level 2→3 instead of 1→2)
and negative progress percentage (-24%). Now all frontend pages compute
level from cumulative_xp via getLevelForXp(), and a migration fixes
stale DB values.

+++ services/pcs-sync/scoring.py
@@ -27,7 +27,7 @@ CONVERSION_RATE = int(os.getenv("CONVERSION_RATE_EUR_PER_PCS", "1500"))
 LEVEL_THRESHOLDS = [0, 50, 150, 300, 500, 700, 1000, 1400, 1900, 2500]
 
 
-def compute_level(xp: int) -> int:
+def compute_level(xp: float) -> int:
     """Compute team level from cumulative XP."""
     for i in range(len(LEVEL_THRESHOLDS) - 1, -1, -1):
         if xp >= LEVEL_THRESHOLDS[i]:
